limit inference specialties to training specialties
